+++ src/back/Plants.Application/Commands/EditPlantCommand.cs
@@ -2,8 +2,8 @@
 
 namespace Plants.Application.Commands;
 
-public record EditPlantCommand(int PlantId, string PlantName,
-    string PlantDescription, int[] RegionIds, int SoilId, int GroupId,
-    int [] RemovedImages, byte[][] NewImages) : IRequest<EditPlantResult>;
+public record EditPlantCommand(long PlantId, string PlantName,
+    string PlantDescription, long[] RegionIds, long SoilId, long GroupId,
+    long[] RemovedImages, byte[][] NewImages) : IRequest<EditPlantResult>;
 
 public record EditPlantResult();
